Fixed plugin and made matrix explicit 32floats.

+++ Cura/Math/Matrix.py
@@ -5,7 +5,7 @@ import math
 # This class is a 4x4 homogenous matrix wrapper arround numpy. 
 
 class Matrix(object):
-    def __init__(self, data = numpy.identity(4)):
+    def __init__(self, data = numpy.identity(4,dtype=numpy.float32)):
         self._data = data
     
     def at(x,y): #TODO add out of index checking
@@ -24,16 +24,16 @@ class Matrix(object):
     
     ## Create a 4x4 identity matrix. This overwrites any existing data.
     def setToIdentity(self):
-        self._data = numpy.identity(4)
+        self._data = numpy.identity(4,dtype=numpy.float32)
         
     ## Invert the matrix
     def invert(self):
-        self._data = numpy.linalg.inv(self._data)
+        self._data = numpy.linalg.inv(self._data,dtype=numpy.float32)
     
     ## Return a inverted copy of the matrix.
     # \returns The invertex matrix.
     def getInverse(self):
-        return Matrix(numpy.linalg.inv(self._data))
+        return Matrix(numpy.linalg.inv(self._data,dtype=numpy.float32))
     
     ## Translate the matrix based on Vector.
     # \param direction The vector by which the matrix needs to be translated.
@@ -45,7 +45,7 @@ class Matrix(object):
     ## Set the matrix by translation vector. This overwrites any existing data.
     # \param direction The vector by which the (unit) matrix needs to be translated.
     def setByTranslation(self, direction):
-        M = numpy.identity(4)
+        M = numpy.identity(4,dtype=numpy.float32)
         M[:3, 3] = direction.getData()[:3]
         self._data = M
